requirements.txt
# RF-DETR Vehicle Detection Dependencies
# Vehicle Detection & Tracking
rfdetr==1.5.0
supervision==0.27.0.post1

# Deep Learning
torch==2.10.0
torchvision==0.25.0

# ONNX Runtime (use onnxruntime for CPU-only, onnxruntime-gpu for GPU)
onnxruntime==1.23.2
# onnxruntime-gpu==1.23.2  # Uncomment for GPU support
onnx==1.19.1

# OCR
paddlepaddle==2.6.2
# paddlepaddle-gpu==2.6.2  # Uncomment for GPU support
paddleocr==2.7.0.3

# Image Processing
opencv-python==4.6.0.66
Pillow==12.1.1
numpy==1.26.4

# Utilities
tqdm==4.67.3
PyYAML==6.0.2  # Use onnxruntime for CPU-only
